Moved ODometry file subsytem to util
