+++ src/main/java/frc/robot/util/Odometry.java
@@ -5,7 +5,7 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.subsystems;
+package frc.robot.util;
 
 import edu.wpi.first.networktables.NetworkTableEntry;
 import edu.wpi.first.networktables.NetworkTableInstance;
@@ -16,8 +16,9 @@ import edu.wpi.first.wpilibj.kinematics.DifferentialDriveOdometry;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Constants;
 import frc.robot.Variables;
+import frc.robot.subsystems.TurretCameraAim;
 
-public class Odometry extends SubsystemBase {
+public class Odometry  {
   double v;
   double lastPosition;
   double currentAngle;
@@ -128,8 +129,5 @@ public class Odometry extends SubsystemBase {
 
   }
 
-  @Override
-  public void periodic() {
-    // This method will be called once per scheduler run
-  }
+  
 }
